Fix for abstract sockets closing.

+++ extensions/localsocket/src/jolie/net/LocalSocketListener.java
@@ -55,7 +55,9 @@ public class LocalSocketListener extends CommListener
 	@Override
 	public void shutdown()
 	{
-		new File( socketAddress.getPath() ).delete();
+		if ( !socketAddress.isAbstract() ) {
+			new File( socketAddress.getPath() ).delete();
+		}
 	}
 
 	@Override
